+++ client/src/frontend/Utils/AppHooks.ts
@@ -1,8 +1,35 @@
-import { useCallback, useState } from 'react';
-import { Wrapper } from '../../backend/Utils/Wrapper';
+import { useCallback, useState, useEffect } from 'react';
 import type GameManager from '../../backend/GameManager';
-import { useEmitterSubscribe } from './EmitterHooks';
+import type { Callback, Monomitter } from '@darkforest_eth/events';
 import type { CommitmentInfo, Tile } from '../../utils';
+
+/**
+ * React uses referential identity to detect changes, and rerender. Rather
+ * than copying an object into a new object, to force a rerender, we can
+ * just wrap it in a new {@code Wrapper}, which will force a rerender.
+ */
+class Wrapper<T> {
+  public readonly value: T;
+
+  public constructor(value: T) {
+    this.value = value;
+  }
+
+  public or(wrapper: Wrapper<T>) {
+    return new Wrapper(this.value || wrapper.value);
+  }
+}
+
+/**
+ * Execute something on emitter callback
+ * @param emitter `Monomitter` to subscribe to
+ * @param callback callback to subscribe
+ */
+function useEmitterSubscribe<T>(emitter: Monomitter<T> | undefined, callback: Callback<T>) {
+  useEffect(() => {
+    return emitter?.subscribe(callback).unsubscribe;
+  }, [emitter, callback]);
+}
 
 export function useTiles(gameManager: GameManager | undefined): Wrapper<Tile[][]> {
   const [tiles, setTiles] = useState<Wrapper<Tile[][]>>(
